dashboard/requirements.txt
Flask
Werkzeug
gunicorn
pymongo
python-dotenv
pyotp
qrcode
pillow
Flask-WTF
Flask-Cors